refactor(errorboundary): move ErrorBoundary component to src/components/common
